Create Info on the project
Implemented a Python-based AI to model Bayesian Networks in order to predict genetic traits. The AI takes in family data and uses probabilistic inference to compute the likelihood that each family member possesses a specific gene and/or exhibits a certain trait. The focus of the project is on the GJB2 gene, which is a leading cause of hearing impairment in newborns. The model utilizes functions to calculate joint probabilities, update those probabilities, and normalize the results. Built for Python 3.10, the project is part of a course that explores applications of artificial intelligence